templates/common/pyproject.toml
[project]
name = "game-mcp-server"
version = "1.0.0"
description = "MCP task management server for {{GAME_NAME}}"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.25,<2",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "Pillow>=12.0",
    "rembg[cpu]>=2.0.73",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/game_mcp"]
